Move copied types to compat/linux/types.h
